Added Site news to the site.
Includes showing the most recent site news story to the home page.

Also noticed that there are enough session summaries on the sessions page that it was annoying to scroll down to see the most recent one. I added an automatic scroll into view on page load function to the page.

Tweaked the visuals on the sessions page a little.